Add error handling
The code should handle exceptions if the URL is incorrect or the website is not accessible. With this updated code, if an error occurs while making a request to the URL or if the response status code is not 200, the scrape_web function will raise an exception and print an error message.